Projects/python/demo_docker/requirements.txt
Django==2.2.7
psycopg2-binary==2.8.4
django-jet==1.0.8
djangorestframework==3.11.0
pyjwt==1.7.1
celery==4.4.1
redis==3.4.1
